# apps/api-gateway/docs/openapi.json
{
  "components": {
    "responses": {
      "ForbiddenError": {
        "content": {
          "application/json": {
            "example": {
              "detail": "admin role required"
            },
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "User terautentikasi tetapi role-nya tidak memiliki akses."
      },
      "UnauthorizedError": {
        "content": {
          "application/json": {
            "example": {
              "detail": "Tidak terautentikasi"
            },
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Firebase ID token tidak ada, tidak valid, atau kedaluwarsa."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "description": "Pesan error yang aman ditampilkan kepada client.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "type": "object"
                },
                "type": "array"
              }
            ]
          }
        },
        "required": [
          "detail"
        ],
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModelRegisterRequest": {
        "properties": {
          "artifact_path": {
            "title": "Artifact Path",
            "type": "string"
          },
          "feature_set_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feature Set Version"
          },
          "horizon": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Horizon"
          },
          "is_active": {
            "default": false,
            "title": "Is Active",
            "type": "boolean"
          },
          "metrics": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "model_type": {
            "title": "Model Type",
            "type": "string"
          },
          "params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params"
          },
          "target_type": {
            "title": "Target Type",
            "type": "string"
          },
          "training_end_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Training End Date"
          },
          "training_start_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Training Start Date"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "model_name",
          "model_type",
          "target_type",
          "version",
          "artifact_path"
        ],
        "title": "ModelRegisterRequest",
        "type": "object"
      },
      "ModelRegistryRow": {
        "properties": {
          "artifact_path": {
            "title": "Artifact Path",
            "type": "string"
          },
          "feature_set_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feature Set Version"
          },
          "horizon": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Horizon"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "metrics": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "model_type": {
            "title": "Model Type",
            "type": "string"
          },
          "params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params"
          },
          "target_type": {
            "title": "Target Type",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "model_name",
          "model_type",
          "target_type",
          "version",
          "artifact_path",
          "is_active"
        ],
        "title": "ModelRegistryRow",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "FirebaseBearer": {
        "bearerFormat": "JWT",
        "description": "Firebase ID token. Masukkan token saja; Swagger UI otomatis menambahkan prefix `Bearer`.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "team@inflasi.id",
      "name": "INFLASI Team",
      "url": "https://inflasi.id/"
    },
...
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "summary": "Food-inflation monitoring + forecasting backend (FastAPI + TimescaleDB + ml-gateway).",
    "title": "INFLASI API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/admin/models": {
      "get": {
        "operationId": "list_models_api_admin_models_get",
        "parameters": [
          {
            "in": "query",
            "name": "model_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Type"
            }
          },
          {
            "in": "query",
            "name": "target_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target Type"
            }
          },
          {
            "in": "query",
            "name": "horizon",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Horizon"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ModelRegistryRow"
                  },
                  "title": "Response List Models Api Admin Models Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "FirebaseBearer": []
          }
        ],
        "summary": "List Models",
        "tags": [
          "admin-models"
        ]
      },
      "post": {
        "operationId": "register_model_api_admin_models_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelRegistryRow"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "FirebaseBearer": []
          }
        ],
        "summary": "Register Model",
        "tags": [
          "admin-models"
        ]
      }
    },
    "/api/admin/models/active": {
      "get": {
        "operationId": "list_active_api_admin_models_active_get",
        "parameters": [
          {
            "in": "query",
            "name": "model_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Type"
            }
          },
          {
            "in": "query",
            "name": "target_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target Type"
            }
          },
          {
            "in": "query",
            "name": "horizon",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Horizon"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ModelRegistryRow"
                  },
                  "title": "Response List Active Api Admin Models Active Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "FirebaseBearer": []
          }
        ],
        "summary": "List Active",
        "tags": [
          "admin-models"
        ]
      }
    },
...
    "/api/internal/models": {
      "post": {
        "description": "Register an artifact produced by a Kubernetes training Job.",
        "operationId": "register_model_from_training_api_internal_models_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Training-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Training-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelRegistryRow"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Model From Training",
        "tags": [
          "internal"
        ]
      }
    },
    "/api/internal/models/active": {
      "get": {
        "description": "Return the single active model row for the given slot.\n\nReturns 404 when no active model is registered; callers should treat that as\n\"this model isn't available yet\" and degrade gracefully.",
        "operationId": "get_active_model_api_internal_models_active_get",
        "parameters": [
          {
            "description": "lightgbm | prophet | sarimax | tft | ensemble | stacking",
            "in": "query",
            "name": "model_type",
            "required": true,
            "schema": {
              "description": "lightgbm | prophet | sarimax | tft | ensemble | stacking",
              "title": "Model Type",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "target_type",
            "required": false,
            "schema": {
              "default": "price",
              "title": "Target Type",
              "type": "string"
            }
          },
          {
            "description": "Required for horizon-scoped models (lightgbm/tft)",
            "in": "query",
            "name": "horizon",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Required for horizon-scoped models (lightgbm/tft)",
              "title": "Horizon"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Active Model Api Internal Models Active Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Active Model",
        "tags": [
          "internal"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Server saat ini (relative)",
      "url": "/"
    },
    {
      "description": "Local dev",
      "url": "http://localhost:8000"
    },
    {
      "description": "Production",
      "url": "https://api.inflasi.id"
    }
  ],
...
}